Move GraphQL queries to WorkflowService (closes #346)
