+++ src/main.js
@@ -17,19 +17,12 @@ import store from '@/store'
 
 // GraphQL client
 import SubscriptionWorkflowService from 'workflow-service'
-import { createApolloClient } from '@/utils/graphql'
 
 // Sync store with router
 sync(store, router)
 
-// TODO: revisit this and evaluate other ways to build the GraphQL URL - not safe to rely on window.location (?)
-const baseUrl = `${window.location.hostname}${window.location.port ? ':' + window.location.port : ''}${window.location.pathname}`
-const httpUrl = `${window.location.protocol}//${baseUrl}graphql`
-const wsUrl = `ws://${baseUrl}subscriptions`
-Vue.prototype.$apolloClient = createApolloClient(httpUrl, wsUrl)
-
 // WorkflowService singleton available application-wide
-const workflowService = new SubscriptionWorkflowService(Vue.prototype.$apolloClient)
+const workflowService = new SubscriptionWorkflowService()
 Vue.prototype.$workflowService = workflowService
 
 Vue.config.productionTip = false
